Fix Presets Scroll; Fix Placeholder position in the quantity field; Hide rendering error in chat; (#555)

+++ src/main/java/codechicken/nei/ItemHistoryPanel.java
@@ -69,10 +69,10 @@ public class ItemHistoryPanel extends Widget {
 
     private void drawSplittingArea(int x, int y, int width, int height, int color) {
 
-        float alpha = (float) (color >> 24 & 255) / 255.0F;
-        float red = (float) (color >> 16 & 255) / 255.0F;
-        float green = (float) (color >> 8 & 255) / 255.0F;
-        float blue = (float) (color & 255) / 255.0F;
+        float alpha = (color >> 24 & 255) / 255.0F;
+        float red = (color >> 16 & 255) / 255.0F;
+        float green = (color >> 8 & 255) / 255.0F;
+        float blue = (color & 255) / 255.0F;
 
         GL11.glPushMatrix();
